moved createObserver method from service to view
