+++ public/js/service-common.js
@@ -1,3 +1,5 @@
+import { ObserversView } from "./view-observer.js";
+
 export class CommonService {
   /**
    * Method used to return error string from the input error response object
@@ -36,41 +38,6 @@ export class CommonService {
   }
 
   /**
-   * Method used to create a observer from the provided HTML element
-   * @param {Element} observerHtmlElement The HTML content from which we want to create an observer object
-   * @returns Object with observer data retrieved from input HTML element
-   */
-  static createObserver(observerHtmlElement) {
-    // we need to check image auxiliary value to correctly determine if empty or not
-    const imageAux = observerHtmlElement.querySelector("input.component-image-auxiliary").value;
-    return {
-      name: observerHtmlElement.querySelector("input.observer-name").value,
-      path: observerHtmlElement.querySelector("input.observer-path").value,
-      target: observerHtmlElement.querySelector("select.observer-target").value,
-      history: observerHtmlElement.querySelector("select.observer-history").value,
-      container: observerHtmlElement.querySelector("input.observer-container").value,
-      title: {
-        interval: "",
-        selector: observerHtmlElement.querySelector("input.component-title-selector").value,
-        attribute: observerHtmlElement.querySelector("input.component-title-attribute").value,
-        auxiliary: observerHtmlElement.querySelector("input.component-title-auxiliary").value,
-      },
-      image: {
-        interval: "",
-        selector: observerHtmlElement.querySelector("input.component-image-selector").value,
-        attribute: observerHtmlElement.querySelector("input.component-image-attribute").value,
-        auxiliary: imageAux === "Select image" ? "" : imageAux,
-      },
-      price: {
-        interval: "",
-        selector: observerHtmlElement.querySelector("input.component-price-selector").value,
-        attribute: observerHtmlElement.querySelector("input.component-price-attribute").value,
-        auxiliary: observerHtmlElement.querySelector("select.component-price-auxiliary").value,
-      },
-    };
-  }
-
-  /**
    * Method used to create an array of observer objects from provided list of HTML elements
    * @param {NodeList} observers The list of HTML observer elemenets
    * @returns array of observer objects created from provided input
@@ -79,7 +46,7 @@ export class CommonService {
     const result = [];
     observers.forEach((observer) => {
       const observerContent = observer.parentNode.querySelector("div.modal-content");
-      result.push(this.createObserver(observerContent));
+      result.push(ObserversView.createObserver(observerContent));
     });
     return result;
   }
