+++ public/js/service-observer.js
@@ -1,4 +1,5 @@
 import { CommonService } from "./service-common.js";
+import { ObserversView } from "./view-observer.js";
 
 export class ObserversService {
   /**
@@ -77,7 +78,7 @@ export class ObserversService {
   static #createRequestOptions(method) {
     const shouldHaveBody = "POST" === method || "PUT" === method;
     const openedObserver = document.querySelector("div.modal-dialog.init-reveal:not(.hidden)");
-    const requestBody = shouldHaveBody ? JSON.stringify(CommonService.createObserver(openedObserver)) : undefined;
+    const requestBody = shouldHaveBody ? JSON.stringify(ObserversView.createObserver(openedObserver)) : undefined;
     return {
       method: method,
       headers: { "Content-Type": "application/json" },
